+++ eventsourcing-postgresql/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Removed Pipes.Consumer from helper functions.
 * Add `TrackedSQLQuery` to run a SQL query while fetching the state from a
   tracking table at the same time. This can be useful with `TopUpReadModel`.
+* Add helpers for migrations: `lockTable` and `swapTables`.
 
 ## 0.9.0 -- 2020-08-16